requirements.txt
rich
pyyaml
h5py
numpy